Convert last_on_time to switch_off_time
This change tracks the last time the heating has been switched off
in order the avoid continuous update to the last_on_time.

In addtition, the swtich_*() methods of the ScriptHeating don't do
any check to the is_on attribute but directly execute the appropriate
switch script. The is_on() method should be checked by the deamon.