Change TimeTable.__setstate__() to transactional
The __setstate__ of TimeTable class now uses the @transactional decorator
and avoids manual storing of old state.

The do_POST() method of the control socket uses the memento function
to revert to old state in case of unsuccessfull update.

Add the status code 503 from control socket when the new settings are
valid but they cannot be saved to filesystem. In that situation the
running daemon continue to run with the new settings but on its restart
the old settings are restored.