Fix bug for unhandled errors in GET /heating requests
During processing the GET /heating requests, the socket queries the
heating and the thermometer. These actions can raise exceptions in
case of comunication failures or other script-related errors.

Now those exceptions are managed to maintain the daemon in a safe state.
The reported error messages need to be checked again.