Fix deepcopy for lock objects in Thread class
During a memento of a Timetable object, also _heating and _thermometer
are deep-copied, but if _thermometer is an instance of
PiAnalogZeroThermometer it has a running thread that cannot be
deep-copied easily. So _heating and _thermometer are now excluded
from memento transaction because they cannot be cheanged neither
remotely nor using config file.