Add custom MCP3008 implementation with spidev module
Importing gpiozero module is very slow in a Raspberry Pi, so a custom
implementation of MCP3008 class is added using only spidev module
to speed up thermod loading. If spidev is not present, the base class
from gpiozero is used with its software spidev.

Moved modules import for Raspberry Pi at the beggining of heating.py
and thermomemter.py files.