Add draft to use proxy in webservers from web manager
Since now the PHP web manager will use cURL to contact Thermod socket
in order to not expose that socket to the outside world. Now the
proxy function of webservers can be used to achive the same masking.

Some adjustments need to be done to support direct error status code
from Thermod socket because previously that error code was transformed
by cURL in the JSON field socket_http_code.