Change startup parameters and log file
If started without parameters, the daemon logs messages to stdout. To have
log printed to syslog, the daemon must be started with '--daemon' option.
If '--daemon' and '--log' and '--debug' options are present, debug and info
messages will be printed only to log file, while warning and errors will
be printed both to syslog and log file.